requirements_base.txt
hydra-core
Pillow
pandas
numpy
scikit-learn
matplotlib
seaborn
pytorch-gradcam
pytorch_lightning